Add equals operator.
