+++ docs/tests/test-equals-operator.yml
@@ -1,0 +1,40 @@
+data:
+  Domain-Seeds:
+    - domain: example.com
+  Expected-IPs:
+    - ip: 2606:2800:220:1:248:1893:25c8:1946
+    - ip: 93.184.216.34
+  Expected-DNS-Entries:
+    - ip: 2606:2800:220:1:248:1893:25c8:1946
+      domain: example.com
+    - ip: 93.184.216.34
+      domain: example.com
+  More-DNS-Entries:
+    - ip: 2606:2800:220:1:248:1893:25c8:1946
+      domain: example.com
+    - ip: 93.184.216.34
+      domain: example.com
+    - ip: 8.8.8.8
+      domain: example.com
+  Less-DNS-Entries:
+    - ip: 2606:2800:220:1:248:1893:25c8:1946
+      domain: example.com
+
+run:
+  - discover Domains and IPs:
+      seeds: use Domain-Seeds
+    find:
+      - IPs
+      - Domains
+      - DNS-Entries
+    expect:
+      - Domains equals Expected-DNS-Entries, otherwise alert high
+      - IPs equals Expected-IPs, otherwise alert high
+      - Expected-DNS-Entries equals DNS-Entries, otherwise alert high
+      - More-DNS-Entries not equals DNS-Entries, otherwise alert high
+      - Less-DNS-Entries not equals DNS-Entries, otherwise alert high
+      
+output:
+  - Template:
+      filename: report-test-equals-operator.html
+      template: templates/html/main.j2
